+++ source/global/management/src/G4AllocatorPool.cc
@@ -21,7 +21,7 @@
 // ********************************************************************
 //
 //
-// $Id: G4AllocatorPool.cc,v 1.1 2004-05-26 14:38:06 gcosmo Exp $
+// $Id: G4AllocatorPool.cc,v 1.2 2004-11-12 16:25:34 gcosmo Exp $
 // GEANT4 tag $Name: not supported by cvs2svn $
 //
 // 
@@ -39,7 +39,7 @@
 // G4AllocatorPool constructor
 // ************************************************************
 //
-G4AllocatorPool::G4AllocatorPool( size_t sz )
+G4AllocatorPool::G4AllocatorPool( unsigned int sz )
   : esize(sz<sizeof(G4PoolLink) ? sizeof(G4PoolLink) : sz),
     csize(sz<1024/2-16 ? 1024-16 : sz*10-16)
 {
